pyproject.toml
[project]
name = "sdg0kubeflow-pipeline-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kfp>=2.16.0",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "minio>=7.2.20",
    "psycopg2-binary>=2.9.12",
    "pymilvus>=2.6.12",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
